requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.2.0
Django==4.2
django-extra-fields==3.0.2
django-os-geocoder==0.1.6
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
phonenumbers==8.13.21
Pillow==9.5.0
progressbar==2.5
psycopg2==2.9.6
PyJWT==2.6.0
pytz==2023.3
razorpay==1.3.0
requests==2.31.0
South==1.0.2
sqlparse==0.4.3
tzdata==2023.3
urllib3==2.0.3
